create new catalog entry for WhitePapers
created new catalog entry with an altered docker compose file. the
compose file is based on the 'sequencing' compose file, but it has
ulimit set for stack; the original sequencing setup was running into
stack limit errors when a script used to generate random forests was
used. adjusting this value should eliminate that.